Port to native macOS ARM64 (Apple Silicon)
- Replace Alure2/OpenAL audio with miniaudio (single-header, Core Audio native)
- Replace GLEW with GLAD (GL 2.1 compatibility profile)
- Modernize CMakeLists.txt: use find_package() instead of hardcoded paths
- Fix C++17 narrowing conversions in Point.h, Range.h, Combat.cpp, etc.
- Add missing #include <string> for modern libc++ compatibility
- Update ASIO to modern API (io_context, resolve without query)
- Set default server to Cosmic (192.168.0.25:8484)
- Trim NxFiles to v83 assets (remove v167+/v209+ entries)
- Rename project to cosmic-client

+++ Gameplay/Combat/DamageNumber.cpp
@@ -21,6 +21,7 @@
 
 #include <nlnx/nx.hpp>
 #include <nlnx/node.hpp>
+#include <string>
 
 namespace ms
 {
@@ -101,7 +102,7 @@ namespace ms
 				for (size_t i = 0; i < restnum.length(); i++)
 				{
 					char c = restnum[i];
-					Point<int16_t> yshift = {0, (i % 2) ? -2 : 2};
+					Point<int16_t> yshift = {static_cast<int16_t>(0), static_cast<int16_t>((i % 2) ? -2 : 2)};
 					charsets[type][true].draw(c, {position + yshift, interopc});
 
 					int16_t advance;
